Refactoring after rebase
