+++ secure-api-gateway-ob-uk-common-datamodel/src/main/java/com/forgerock/sapi/gateway/ob/uk/common/datamodel/account/FRAccountServicer.java
@@ -41,5 +41,6 @@ import lombok.NoArgsConstructor;
 public class FRAccountServicer {
 
     private String schemeName;
+    private String name;
     private String identification;
 }
