+++ secure-api-gateway-ob-uk-common-datamodel/src/main/java/com/forgerock/sapi/gateway/ob/uk/common/datamodel/account/FRPartyData.java
@@ -52,6 +52,7 @@ public class FRPartyData {
     private String mobile;
     private FRRelationship relationship;
     private List<FRPostalAddress> addresses;
+    private String LEI;
 
     public enum FRPartyType {
         DELEGATE("Delegate"),
