+++ secure-api-gateway-ob-uk-common-datamodel/src/main/java/com/forgerock/sapi/gateway/ob/uk/common/datamodel/converter/v4/account/FREntryStatusConverter.java
@@ -13,9 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.forgerock.sapi.gateway.ob.uk.common.datamodel.v4.converter.account;
+package com.forgerock.sapi.gateway.ob.uk.common.datamodel.converter.v4.account;
 
-import com.forgerock.sapi.gateway.ob.uk.common.datamodel.v4.account.FRTransactionData.FREntryStatus;
+import com.forgerock.sapi.gateway.ob.uk.common.datamodel.account.FRTransactionData.FREntryStatus;
 import uk.org.openbanking.datamodel.v4.account.ExternalEntryStatus1Code;
 
 import java.util.Collections;
