+++ secure-api-gateway-ob-uk-common-datamodel/src/main/java/com/forgerock/sapi/gateway/ob/uk/common/datamodel/converter/v4/account/FRExternalMandateStatusConverter.java
@@ -13,10 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.forgerock.sapi.gateway.ob.uk.common.datamodel.v4.converter.account;
+package com.forgerock.sapi.gateway.ob.uk.common.datamodel.converter.v4.account;
 
 import com.forgerock.sapi.gateway.ob.uk.common.datamodel.v4.account.FRDirectDebitData.FRDirectDebitStatus;
-import com.forgerock.sapi.gateway.ob.uk.common.datamodel.v4.account.FRStandingOrderData;
+import com.forgerock.sapi.gateway.ob.uk.common.datamodel.account.FRStandingOrderData;
 import uk.org.openbanking.datamodel.v4.account.ExternalMandateStatus1Code;
 
 import java.util.Collections;
