+++ secure-api-gateway-ob-uk-common-datamodel/src/main/java/com/forgerock/sapi/gateway/ob/uk/common/datamodel/payment/FRExternalPaymentPurposeCode.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.forgerock.sapi.gateway.ob.uk.common.datamodel.v4.payment;
+package com.forgerock.sapi.gateway.ob.uk.common.datamodel.payment;
 
 import lombok.Getter;
